# post-data-management-back-end-development/docs/swagger.yaml
basePath: /v1
definitions:
  apierrors.APIErrorResponse:
    properties:
      error:
        $ref: '#/definitions/apierrors.AppError'
      message:
        description: Descriptive message providing details about the error.
        type: string
      status_code:
        description: HTTP status code representing the error type (e.g., 404 for Not
          Found).
        type: integer
      success:
        description: Success flag to indicate if the operation was successful.
        type: boolean
    type: object
  apierrors.AppError:
    properties:
      code:
        type: string
      field_errors:
        items:
          $ref: '#/definitions/apierrors.FieldError'
        type: array
      id:
        type: string
      message:
        type: string
    type: object
  apierrors.FieldError:
    properties:
      field:
        type: string
      message:
        type: string
      value: {}
    type: object
...
  handler.CreateCadreRequest:
    properties:
      cadre_name:
        type: string
      created_by:
        type: string
      grade_pay:
        type: integer
      group_id:
        type: integer
      group_name:
        type: string
      pay_level:
        type: integer
      remarks:
        type: string
      status:
        type: string
      valid_from:
        type: string
      valid_to:
        type: string
    required:
    - cadre_name
    - created_by
    - grade_pay
    - group_id
    - group_name
    - pay_level
    - remarks
    - status
    - valid_from
    - valid_to
    type: object
...
  handler.UpdateCadreRequest:
    properties:
      cadre_name:
        $ref: '#/definitions/null.String'
      cadreID:
        $ref: '#/definitions/null.Int32'
      grade_pay:
        $ref: '#/definitions/null.Int32'
      group_id:
        $ref: '#/definitions/null.Int16'
      group_name:
        $ref: '#/definitions/null.String'
      pay_level:
        $ref: '#/definitions/null.Int32'
      remarks:
        $ref: '#/definitions/null.String'
      status:
        $ref: '#/definitions/null.String'
      updated_by:
        $ref: '#/definitions/null.String'
      valid_from:
        type: string
      valid_to:
        type: string
    required:
    - cadreID
    - cadre_name
    - grade_pay
    - group_id
    - group_name
    - pay_level
    - remarks
    - status
    - updated_by
    - valid_from
    - valid_to
    type: object
...
  null.Int16:
    properties:
      int16:
        type: integer
      set:
        type: boolean
      valid:
        type: boolean
    type: object
  null.Int32:
    properties:
      int32:
        type: integer
      set:
        type: boolean
      valid:
        type: boolean
    type: object
  null.String:
    properties:
      set:
        type: boolean
      string:
        type: string
      valid:
        type: boolean
    type: object
...
  response.CreateCadreMasterAPIResponse:
    properties:
      data:
        $ref: '#/definitions/response.CreateCadreMasterResponse'
      message:
        type: string
      status_code:
        type: integer
      success:
        type: boolean
    type: object
  response.CreateCadreMasterResponse:
    properties:
      cadre_id:
        type: integer
    type: object
...
  response.GetPostNameMasterAPIResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/response.GetPostNameMasterResponse'
        type: array
      limit:
        type: integer
      message:
        type: string
      order_by:
        type: string
      returned_records_count:
        type: integer
      skip:
        type: integer
      sort_type:
        type: string
      status_code:
        type: integer
      success:
        type: boolean
      total_records_count:
        type: integer
    type: object
  response.GetPostNameMasterResponse:
    properties:
      cadre:
        type: string
      group:
        type: string
      post_name:
        type: string
    type: object
...
  response.ListCadresAPIResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/response.ListCadresResponse'
        type: array
      limit:
        type: integer
      message:
        type: string
      order_by:
        type: string
      returned_records_count:
        type: integer
      skip:
        type: integer
      sort_type:
        type: string
      status_code:
        type: integer
      success:
        type: boolean
      total_records_count:
        type: integer
    type: object
  response.ListCadresResponse:
    properties:
      cadre_id:
        type: integer
      cadre_name:
        type: string
      group_name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support_cept@indiapost.gov.in
    name: API Support Team
    url: http://cept.gov.in/support
  description: A comprehensive API for managing Post Management Data Management, offering
    endpoints for creation, update, deletion, and retrieval of Posts Data.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://cept.gov.in/terms
  title: Post Management Data Management API
  version: "1.0"
paths:
...
  /post-management/cadres:
    get:
      consumes:
      - application/json
      description: List all cadre types or filter by group-code if provided
      parameters:
      - description: Filter cadres by group code
        in: query
        name: group-code
        type: string
      - in: query
        name: limit
        type: integer
      - in: query
        name: order_by
        type: string
      - in: query
        name: skip
        type: integer
      - in: query
        name: sort_type
        type: string
      - in: query
        name: total_records_required
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Cadre list retrieved successfully
          schema:
            $ref: '#/definitions/response.ListCadresAPIResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "403":
          description: Forbidden error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "409":
          description: Data conflict error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
      summary: List Cadre Details
      tags:
      - Post Management
  /post-management/cadres/{cadre-id}:
    get:
      consumes:
      - application/json
      description: Fetches post names based on the provided Cadre ID
      parameters:
      - description: Cadre ID
        in: path
        name: cadre-id
        required: true
        type: string
      - in: query
        name: limit
        type: integer
      - in: query
        name: order_by
        type: string
      - in: query
        name: skip
        type: integer
      - in: query
        name: sort_type
        type: string
      - in: query
        name: total_records_required
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Post names retrieved successfully
          schema:
            $ref: '#/definitions/response.GetPostNameMasterAPIResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "403":
          description: Forbidden error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "409":
          description: Data conflict error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
      summary: Get Post Names by Cadre ID
      tags:
      - Post Management
    put:
      consumes:
      - application/json
      description: Updates a post cadre master entry.
      parameters:
      - description: Cadre ID to update
        in: path
        name: cadre-id
        required: true
        type: integer
      - description: cadre details to update
        in: body
        name: UpdateCadreRequest
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateCadreRequest'
      produces:
      - application/json
      responses:
        "200":
          description: resource updated successfully
          schema:
            $ref: '#/definitions/response.CreateCadreMasterAPIResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "401":
          description: Unauthorized access
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "403":
          description: Forbidden access
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "404":
          description: Resource not found
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "409":
          description: Conflict in the data
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
      summary: Update Cadre Master
      tags:
      - Post Management
  /post-management/cadres/all-cadres:
    get:
      consumes:
      - application/json
      description: List all cadre types or filter by group-code if provided
      parameters:
      - description: Filter cadres by group code
        in: query
        name: group-code
        type: string
      - in: query
        name: limit
        type: integer
      - in: query
        name: order_by
        type: string
      - in: query
        name: skip
        type: integer
      - in: query
        name: sort_type
        type: string
      - in: query
        name: total_records_required
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Cadre list retrieved successfully
          schema:
            $ref: '#/definitions/response.ListCadresAPIResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "403":
          description: Forbidden error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "409":
          description: Data conflict error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
      summary: List All Cadre Details
      tags:
      - Post Management
...
  /post-management/cadres/create:
    post:
      consumes:
      - application/json
      description: Creates a post cadre master entry.
      parameters:
      - description: cadre details to create
        in: body
        name: CreateCadreRequest
        required: true
        schema:
          $ref: '#/definitions/handler.CreateCadreRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Cadre master details created successfully
          schema:
            $ref: '#/definitions/response.CreateCadreMasterAPIResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "401":
          description: Unauthorized access
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "403":
          description: Forbidden access
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "404":
          description: Resource not found
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "409":
          description: Conflict in the data
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
      summary: Create Cadre Master
      tags:
      - Post Management
...
schemes:
- http
- https
swagger: "2.0"